Fix blog posts losing data when previewing updates
When editing an existing blog post, and clicking on "Preview" instead of
just saving directly, the blog post would have its date/time updated to
"now" instead of keeping its original time (and regardless of the
checkbox to reset time not being checked).

The fix was to catch the post ID from the submission so it can load the
existing data first; previously it was only getting the post ID from the
query string, which isn't used when previewing the post.